cpu: o3 cpu impl for PIM stats need verify
O3CPU can dump stats for offloaded instruction; Need to find out reason for stats arthmatic count not the same as profiling. This can be due to the first iteration not logged. Need to correct timing simple cpu so that the stats will be updated only when a new inst is encountered. Need to figure out why o3 cpu PIM fraction is not the same as timing simple cpu.

Change-Id: I882fa59aacf1535907be639f8d27726a46e961ad